refactor(api): restructure payment endpoints and types
