+++ shared/types/index.ts
@@ -160,10 +160,7 @@ export declare class MonetizationEvent extends Event {
 
 export const KV_PAYMENTS_PREFIX = 'payments/'
 
-export const PAYMENT_ERROR = {
-  NON_POSITIVE_AMOUNT: 'NON_POSITIVE_AMOUNT',
-} as const
-export type PaymentError = keyof typeof PAYMENT_ERROR
+export type PaymentError = 'NON_POSITIVE_AMOUNT'
 
 export const BANNER_TITLE_MAX_LENGTH = 60
 export const BANNER_DESCRIPTION_MAX_LENGTH = 300
